remove randomHex() from testutil
